Update StorageFile to load file based on new project structure
The load method will first scan each line of String for an event. It
will then construct that event. If the next line of String is a
task, it will add that task to the last constructed event. The main
assertion is that the first line in the save file always has to be
an event.