Use generic interfaces instead of implementations
For example, use List for ArrayList, and Map for HashMap.

This is considered good practice since it enables better maintainability, say, if we want to use LinkedList in place of ArrayList someday.

TreeSet is yet to be replaced since we make use of the first() and last() methods, which should not have been used for a Set ADT in the first place. We may want to factor those out as well some day.